planner-agent-/app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
langchain>=0.0.200
langchain-community>=0.0.10
motor>=3.1.1
pymongo>=4.3.3
python-jose[cryptography]>=3.3.0
openai>=1.0.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
pydantic>=2.0.0
python-multipart>=0.0.5
